requirements.txt
slack-bolt==1.18.1
slack-sdk==3.27.1
anthropic>=0.50.0
httpx>=0.25.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.116.0
pyairtable==2.3.3
python-dotenv==1.0.1
aiohttp==3.9.3
pypdf==4.0.1
requests==2.31.0
tavily-python==0.3.9
pyparsing>=3.0.0
httplib2>=0.22.0
